code/requirements.txt
# Core FastAPI stack
fastapi==0.111.0
uvicorn==0.23.2
starlette==0.37.2
fastapi-cli==0.0.13
jinja2==3.1.2
python-multipart==0.0.7
aiofiles==23.2.1
pydantic==2.11.10
pydantic_core==2.33.2

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
databases==0.9.0
orjson==3.11.3

# HTTP & Requests
requests==2.31.0
httpx==0.28.1
httpcore==1.0.9
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
certifi==2025.10.5
dnspython==2.8.0

# Security / Encryption
cryptography==41.0.3
cffi==2.0.0
colorama==0.4.6
email-validator==2.3.0

# Security for user login/register
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
bcrypt


# Utilities / CLI / Rich UI
rich==14.1.0
rich-toolkit==0.15.1
typer==0.19.2
shellingham==1.5.4
markdown-it-py==4.0.0
pygments==2.19.2
mdurl==0.1.2
python-dotenv==1.1.1
tzdata==2025.2
ujson==5.11.0
typing-inspection==0.4.2
typing-extensions==4.15.0
annotated-types==0.7.0
watchfiles==1.1.0

# Async / Web
websockets==15.0.1
anyio==4.11.0
#uvloop==0.21.0
httptools==0.6.4
h11==0.16.0







